Refactoring for sanity
