+++ downloadArtwork.js
@@ -1,0 +1,12 @@
+const request = require("request");
+const fs = require("fs");
+
+
+module.exports = function download_artwork (uri, filename, callback) {
+    request.head(uri, function (err, res, body) {
+        //console.log('content-type:', res.headers['content-type']);
+        //console.log('content-length:', res.headers['content-length']);
+
+        request(uri).pipe(fs.createWriteStream(filename)).on("close", callback);
+    });
+};
